Bring Quarter logs up-to-date
This does numerous things. it adds a parent json for quarter logs that halves the required textures, redoes the jsons to use the fewer textures and the parent, and adds new textures for the mega trees. this also attempts to dissolve the bald cypress tree, but it still only generates the cypress tree.