Add PredictiveSpruceFoliagePlacer avoiding client freezes / lag
This lag was caused by the leaves updating their distance values, causing a storm of block updates. We fix this by predicting the distance value, therefore significantly reducing the number fo block updates.

This prediction code was present in 1.15 for different reasons (spurious leaf decay), but was removed during the port. It is now restored.