Fix "breaks" bounds using "<=" instead of "<"
Traverse 3.0.0 and Cinderscapes 1.2.3 are both compatible with the current version of Terrestria

Also add a breaks clause for old versions of Campanion. I didn't see any actual crashes with the latest 1.16.1 version, but I know that old versions of Campanion bundled an outdated version of Terraform.

Also, older versions of Campanion aren't compatible with 1.16.2+, so they shouldn't be used with Terrestria for 1.16.2+ anyways.