Update to 1.19.4-pre3.
TerraBlender is not yet available for 1.19.4, so to run this release you will need to download the alpha release of Biolith from:
https://github.com/gniftygnome/Biolith/releases

This is a dev commit checkpointing 1.19.4 changes.  This commit *will* cause discontinuities in biome placement if you use it to upgrade an existing world or if you make a new world using Biolith and then change back to TerraBlender later.  YOU HAVE BEEN WARNED!